moved (un)install to argocd-util
